requirements.txt
asgiref==3.2.7
boto3==1.13.14
botocore==1.16.14
Django==3.0.5
django-filter==2.2.0
django-storages==1.9.1
docutils==0.15.2
jmespath==0.10.0
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
